+++ packages/code-du-travail-frontend/src/modules/outils/indemnite-licenciement/__tests__/absence-date-without-agreement.test.tsx
@@ -16,9 +16,9 @@ describe("Indemnité licenciement - Validation du comportement avec l'absence de
       userAction = new UserAction();
       userAction
         .click(ui.introduction.startButton.get())
-        .click(ui.contract.inaptitude.non.get())
-        .click(ui.next.get())
         .click(ui.agreement.noAgreement.get())
+        .click(ui.next.get())
+        .click(ui.information.inaptitude.non.get())
         .click(ui.next.get());
 
       // Validation que l'on est bien sur l'étape ancienneté