+++ packages/code-du-travail-frontend/src/modules/outils/indemnite-licenciement/__tests__/cc1517.test.tsx
@@ -25,8 +25,8 @@ describe("Indemnité licenciement - CC 1517", () => {
 
     userAction = new UserAction();
     userAction.click(ui.introduction.startButton.get());
-    userAction.click(ui.contract.inaptitude.non.get());
     userAction.click(ui.next.get());
+    userAction.click(ui.information.inaptitude.non.get());
     userAction.click(ui.next.get());
   });
   test("vérifier l'eligibilite pour ancienneté (entree -> sortie) supérieur à 8 mois", async () => {