+++ packages/code-du-travail-frontend/src/modules/outils/indemnite-licenciement/__tests__/cc86.test.tsx
@@ -29,8 +29,9 @@ describe("Indemnité licenciement - CC 86", () => {
 
   test(`Vérification du fait que le salaire de reference est bien celui issu de la formule de calcul de l'indemnité`, () => {
     userAction.click(ui.introduction.startButton.get());
-    userAction.click(ui.contract.inaptitude.non.get());
     userAction.click(ui.next.get());
+    userAction.click(ui.next.get());
+    userAction.click(ui.information.inaptitude.non.get());
     userAction.click(ui.next.get());
     userAction.setInput(ui.seniority.startDate.get(), "01/01/2000");
     userAction.setInput(ui.seniority.notificationDate.get(), "01/01/2022");